added functionality and DB migrations for supporting reviewing stories

+++ src/stories/stories.service.ts
@@ -4,6 +4,7 @@ import {
   InternalServerErrorException,
   Logger,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { Story } from './story.entity';
 import { StoriesRepository } from './stories.repository';
@@ -11,6 +12,9 @@ import { GetStoriesFilterDto } from './dto/get-stories-filter.dto';
 import { CreateStoryDto } from './dto/create-story.dto';
 import { UpdateStoryDto } from './dto/update-story.dto';
 import { GetStories } from './story.types';
+import { User } from 'src/auth/user.entity';
+import { GetStoriesFilterManagerDto } from './dto/get-stories-filter-admin.dto';
+import { Role } from 'src/auth/role.enum';
 
 @Injectable()
 export class StoriesService {
@@ -19,6 +23,10 @@ export class StoriesService {
 
   getStories(filterDto: GetStoriesFilterDto): Promise<GetStories> {
     return this.storiesRepository.getStories(filterDto);
+  }
+
+  getStoriesManager(filterDto: GetStoriesFilterManagerDto): Promise<Story[]> {
+    return this.storiesRepository.getStoriesManager(filterDto);
   }
 
   async getStoryById(id: string): Promise<Story> {
@@ -39,11 +47,20 @@ export class StoriesService {
     }
   }
 
-  createStory(createStoryDto: CreateStoryDto): Promise<Story> {
-    return this.storiesRepository.createStory(createStoryDto);
+  createStory(createStoryDto: CreateStoryDto, user: User): Promise<Story> {
+    return this.storiesRepository.createStory(createStoryDto, user);
   }
 
-  async deleteStoryById(id: string): Promise<void> {
+  async deleteStoryById(id: string, user: User): Promise<void> {
+    const isManager = user.roles.includes(Role.Manager);
+
+    const story = await this.storiesRepository.findOneBy({ id });
+
+    // manager is only allowed to delete story if the story has not been approved yet.
+    if (story.isReviewed && isManager) {
+      throw new UnauthorizedException();
+    }
+
     const result = await this.storiesRepository.delete({ id });
 
     if (result.affected === 0) {
